tdf#86630 sw page number wizard: avoid crash when no GetCurrFrame
The steps to get the crash were:
-create a header (and leave the cursor in the header)
-use the wizard to create a non-mirrored footer

I had seen this crash happening occassionally during development,
but hadn't yet identified the exact steps to consistently reproduce.
Thanks to QA who tracked that down.

Although it looks/feels like a hack, it actually works amazingly well.
The simple act of SetCursorInHdFt "fixes" GetCurrFrame/Get*PageNum,
but we can't always use that because we don't know whether we
are on an odd or even or first page when we start, so Goto*Text
is the best when non-mirroring.

Change-Id: I0e2f28d9c8fd2d1a10afede885b525c42ed08af4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151345
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/151907
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>