xmloff: add color-type attribute to complex color
We need to identify what the color type of the complex color is.
For now we mostly use "theme", but in the future we can also have
other types like "rgb",... when the infrastructure for that is
built.

Change-Id: I38c91d294a191ca3124be4e99050977c9815d23e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/152253
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
(cherry picked from commit 38e0e78998153463caf9c3c72ef7f4549ddff0e8)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/152516
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>