Fix CI breakage from commit 36831c668c53fdfcd78300045e6beafccbba35c6
It backported commit ebf8b0b0699fe34c86badb3087869d902a4e40a8 from
libreoffice-7-5 to co-23.05; and in the process, it manually restored
the 'if (rWrt.mbXHTML)' condition, that was removed in co-23.05 in
commit b401ff5020264adef2fd68c111d9bc9ea84ee1ab (htmlexport: close li
mark not only in xhtml, 2023-02-08).

Unfortunately, the backport that included manual merge conflict
resolution wasn't CI-tested.

Change-Id: Id42b6309ed92e209b682b4b1cf5578e2a0f20fc1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/152385
Reviewed-by: Andras Timar <andras.timar@collabora.com>
Tested-by: Andras Timar <andras.timar@collabora.com>