MCGR: Use tryToRecreateBorder for better BW comp with LO
For better compatibility with LO versions before MCGR, try
to re-create a 'border' value based on the existing GradientSteps.

With MCGR we do not need 'border' anymore in quite some cases since
no Start/EndColor at 0.0 resp. 1.0 is explicitely needed. Since we
(unfortunately need to) internally continue to support border
anyways it does no harm to fallback to use the border value - if
there is an equivalent representation as this helper checks for.

For exports that do not support 'border' this will be adapted as
needed (see tryToApplyBorder())

Change-Id: If98c64039ff97143d4b5c92ac2a950e70f5bb70a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/152395
Tested-by: Jenkins
Reviewed-by: Armin Le Grand <Armin.Le.Grand@me.com>