tdf#153040 sw content controls PDF export: output listbox entry
This fixes a 7.5 regression/wrong implementation.

Prior to 7.5, content controls were just exported as plain text to PDF.
Now that the content control is exported, the selected entry in
the list box was not being printed - just a big hole in the text.

Fixed by informing PDF of the selected entry.

TODO: the content control starts too late
(a somewhat wide space at the front),
and ends too soon (the last portion of the text is cut off).

The unit test for the TODO will cover this case,
so skipping that for now since I found no examples to follow.

Change-Id: I1412a57efc070010c6d0d0a37f48d63699b085b1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/152567
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>